+++ oap-server/server-alarm-plugin/src/main/java/org/apache/skywalking/oap/server/core/alarm/provider/RulesReader.java
@@ -30,6 +30,7 @@ import org.apache.skywalking.oap.server.core.alarm.provider.feishu.FeishuSetting
 import org.apache.skywalking.oap.server.core.alarm.provider.grpc.GRPCAlarmSetting;
 import org.apache.skywalking.oap.server.core.alarm.provider.slack.SlackSettings;
 import org.apache.skywalking.oap.server.core.alarm.provider.wechat.WechatSettings;
+import org.apache.skywalking.oap.server.core.alarm.provider.welink.WeLinkSettings;
 import org.yaml.snakeyaml.Yaml;
 import org.yaml.snakeyaml.constructor.SafeConstructor;
 
@@ -64,6 +65,7 @@ public class RulesReader {
             readCompositeRuleConfig(rules);
             readDingtalkConfig(rules);
             readFeishuConfig(rules);
+            readWeLinkConfig(rules);
         }
         return rules;
     }
@@ -249,6 +251,34 @@ public class RulesReader {
                 });
             }
             rules.setFeishus(feishuSettings);
+        }
+    }
+
+    /**
+     * Read WeLink hook config into {@link WeLinkSettings}
+     */
+    private void readWeLinkConfig(Rules rules) {
+        Map welinkConfig = (Map) yamlData.get("welinkHooks");
+        if (welinkConfig != null) {
+            WeLinkSettings welinkSettings = new WeLinkSettings();
+            Object textTemplate = welinkConfig.getOrDefault("textTemplate", "");
+            welinkSettings.setTextTemplate((String) textTemplate);
+            List<Map<String, Object>> welinkWebHooks = (List<Map<String, Object>>) welinkConfig.get("webhooks");
+            if (welinkWebHooks != null) {
+                welinkWebHooks.forEach(welinkWebhook -> {
+                    String clientId = (String) welinkWebhook.getOrDefault("client_id", "");
+                    String clientSecret = (String) welinkWebhook.getOrDefault("client_secret", "");
+                    String accessTokenUrl = (String) welinkWebhook.getOrDefault("access_token_url", "");
+                    String messageUrl = (String) welinkWebhook.getOrDefault("message_url", "");
+                    String groupIds = (String) welinkWebhook.getOrDefault("group_ids", "");
+                    String rebootName = (String) welinkWebhook.getOrDefault("robot_name", "reboot");
+                    welinkSettings.getWebhooks()
+                                  .add(new WeLinkSettings.WebHookUrl(clientId, clientSecret, accessTokenUrl, messageUrl,
+                                                                     rebootName, groupIds
+                                  ));
+                });
+            }
+            rules.setWelinks(welinkSettings);
         }
     }
 }