+++ CHANGES.md
@@ -23,6 +23,7 @@ Release Notes.
 * Fix K8s monitoring the incorrect metrics calculate. 
 * Loop alarm into event system.
 * Support alarm tags.
+* Support WeLink as a channel of alarm notification.
 
 #### UI
 * Add logo for kong plugin.